Added Backend support for classes and events
Events and classes added are persistent across login sessions now. However, there's a bug where if a previously logged in user had a certain event for a certain class, the next user may see that event even if they are not in the class. RSVP is still a work in progress.